pyproject.toml
[project]
authors = [{ name = "Dan Egnor", email = "egnor@ofb.net" }]
dependencies = [
  "asyncclick>=8.3.0.7",
  "matplotlib>=3.10.8",
  "odrive>=0.6.11.dev1",
  "ok-logging-setup>=0.13",
  "construct>=2.10",
  "fastcrc>=0.3",
  "pyserial>=3.5",
  "python-can>=4.6.1",
  "pyusb>=1.3.1",
  "ok-serial[cli]>=0.3.1",
  "evdev>=1.7",
  "pygame-ce>=2.5.7",
  "pysdl3>=0.9.11b0",
  "ok-subprocess-defaults>=0.2",
  "streamdeck>=0.9.8",
]
description = "Neotropolis project code"
name = "egnor-neotropolis"
readme = "README.md"
requires-python = ">=3.12"
version = "0.0"

[project.scripts]
base_display = "trashbot.base_display_cli:main"
bot_display = "trashbot.bot_display_worker:main"
emoji_input = "trashbot.emoji_input_cli:main"
motor = "trashbot.motor_cli:main"
gamepad = "trashbot.gamepad_cli:main"
radio = "trashbot.radio_cli:main"
run_base = "trashbot.run_base:main"
run_bot = "trashbot.run_bot:main"
system_setup = "trashbot.system_setup_cli:main"

[build-system]
requires = ["uv_build>=0.10.0,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
  "construct-typing",
  "mypy",
  "pytest",
  "pytest-asyncio",
  "pytest-mock",
  "ruff",
]

[tool.mypy]
exclude_gitignore = true
files = "trashbot"
overrides = [
  { module = "odrive.*", follow_untyped_imports = true },
  { module = "StreamDeck.*", follow_untyped_imports = true },
]

[tool.pytest]
asyncio_mode = "auto"
log_level = "CRITICAL"
strict = true

[tool.ruff]
line-length = 80
extend-exclude = ["*_misc"]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]

[tool.uv.build-backend]
module-root = ""
module-name = "trashbot"
